# guanseq-server/src/main/resources/openapi/guanseq-api-v1.yaml
openapi: 3.1.0
info:
  title: GuanSeq API
  version: 0.1.0
  description: GuanSeq 制造业务核心 API 契约。
paths:
...
  /api/v1/finance/receivable-invoices:
    get:
      operationId: listReceivableInvoices
      summary: 分页查询销售应收发票
      tags: [Finance]
      security: [{ basicAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Query"
        - name: status
          in: query
          schema: { type: string, enum: [ALL, OPEN, PARTIALLY_PAID, PAID, CREDIT_PENDING, SETTLED], default: ALL }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: 应收发票分页
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReceivableInvoicePage" }
    post:
      operationId: createReceivableInvoice
      summary: 按已发货销售订单开具应收发票
      tags: [Finance]
      security: [{ basicAuth: [] }]
      parameters:
        - name: X-Request-Id
          in: header
          required: false
          schema: { type: string, maxLength: 120 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateReceivableInvoiceRequest" }
      responses:
        "200":
          description: 已创建或幂等返回应收发票
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReceivableInvoice" }
        "403": { description: 当前角色无权开票 }
        "409": { description: 请求编号或并发冲突 }
        "422": { description: 日期、订单状态或可开票数量不满足约束 }
  /api/v1/finance/receivable-invoices/{id}:
    get:
      operationId: getReceivableInvoice
      summary: 获取应收发票详情及收款记录
      tags: [Finance]
      security: [{ basicAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/EntityId" }]
      responses:
        "200":
          description: 应收发票详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReceivableInvoice" }
        "404": { description: 应收发票不存在或不可见 }
  /api/v1/finance/receivable-invoices/{id}/receipts:
    post:
      operationId: postReceivableReceipt
      summary: 登记收款并核销指定应收发票
      tags: [Finance]
      security: [{ basicAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/EntityId"
        - name: X-Request-Id
          in: header
          required: false
          schema: { type: string, maxLength: 120 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PostReceivableReceiptRequest" }
      responses:
        "200":
          description: 已登记收款并返回更新后的发票
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReceivableInvoice" }
        "403": { description: 当前角色无权登记收款 }
        "409": { description: 发票版本、请求编号或并发冲突 }
        "422": { description: 收款日期或金额不满足约束 }
  /api/v1/finance/receivable-reference-data:
    get:
      operationId: getReceivableReferenceData
      summary: 获取已发货订单及其剩余可开票数量
      tags: [Finance]
      security: [{ basicAuth: [] }]
      responses:
        "200":
          description: 可开票销售订单
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReceivableReferenceData" }
  /api/v1/finance/receivable-credit-notes:
    get:
      operationId: listReceivableCreditNotes
      summary: 分页查询应收红字发票
      tags: [Finance]
      security: [{ basicAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Query"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: 应收红字发票分页
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReceivableCreditNotePage" }
    post:
      operationId: createReceivableCreditNote
      summary: 对已过账应收发票开具红字发票
      tags: [Finance]
      security: [{ basicAuth: [] }]
      parameters:
        - name: X-Request-Id
          in: header
          required: false
          schema: { type: string, maxLength: 120 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateReceivableCreditNoteRequest" }
      responses:
        "200":
          description: 已过账红字发票
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReceivableCreditNote" }
        "403": { description: 当前角色无权开具红字发票 }
        "409": { description: 请求编号或并发冲突 }
        "422": { description: 日期、数量或原发票状态不满足约束 }
  /api/v1/finance/receivable-credit-notes/{id}:
    get:
      operationId: getReceivableCreditNote
      summary: 获取应收红字发票详情
      tags: [Finance]
      security: [{ basicAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/EntityId" }]
      responses:
        "200":
          description: 红字发票详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReceivableCreditNote" }
        "404": { description: 红字发票不存在或不可见 }
  /api/v1/finance/receivable-invoices/{id}/refunds:
    post:
      operationId: postReceivableRefund
      summary: 对存在待退余额的应收发票登记退款
      tags: [Finance]
      security: [{ basicAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/EntityId"
        - name: X-Request-Id
          in: header
          required: false
          schema: { type: string, maxLength: 120 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PostReceivableRefundRequest" }
      responses:
        "200":
          description: 已登记退款并返回更新后的发票
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReceivableInvoice" }
        "403": { description: 当前角色无权登记退款 }
        "409": { description: 发票版本、请求编号或并发冲突 }
        "422": { description: 退款日期、金额或待退余额不满足约束 }
  /api/v1/finance/receivables/receipts/{id}/reverse:
    post:
      operationId: reverseReceivableReceipt
      summary: 反核销一笔应收收款或退款
      tags: [Finance]
      security: [{ basicAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/EntityId"
        - name: X-Request-Id
          in: header
          required: false
          schema: { type: string, maxLength: 120 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReverseReceivableReceiptRequest" }
      responses:
        "200":
          description: 已反核销并返回更新后的发票
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReceivableInvoice" }
        "403": { description: 当前角色无权反核销 }
        "409": { description: 请求编号、重复反核销或并发冲突 }
        "422": { description: 日期或原因不满足约束 }
        "404": { description: 收款记录不存在或不可见 }
  /api/v1/finance/payable-invoices:
    get:
      operationId: listPayableInvoices
      summary: 分页查询采购应付发票
      tags: [Finance]
      security: [{ basicAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Query"
        - name: status
          in: query
          schema: { type: string, enum: [ALL, OPEN, PARTIALLY_PAID, PAID, CREDIT_PENDING, SETTLED], default: ALL }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: 应付发票分页
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PayableInvoicePage" }
    post:
      operationId: createPayableInvoice
      summary: 按采购订单合格收货数量登记应付发票
      tags: [Finance]
      security: [{ basicAuth: [] }]
      parameters:
        - name: X-Request-Id
          in: header
          required: false
          schema: { type: string, maxLength: 120 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePayableInvoiceRequest" }
      responses:
        "200":
          description: 已创建或幂等返回应付发票
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PayableInvoice" }
        "403": { description: 当前角色无权登记应付发票 }
        "409": { description: 请求编号、供应商发票号或并发冲突 }
        "422": { description: 日期、合格收货数量或订单归属不满足约束 }
  /api/v1/finance/payable-invoices/{id}:
    get:
      operationId: getPayableInvoice
      summary: 获取应付发票详情及付款记录
      tags: [Finance]
      security: [{ basicAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/EntityId" }]
      responses:
        "200":
          description: 应付发票详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PayableInvoice" }
        "404": { description: 应付发票不存在或不可见 }
  /api/v1/finance/payable-invoices/{id}/payments:
    post:
      operationId: postPayablePayment
      summary: 登记付款并核销指定应付发票
      tags: [Finance]
      security: [{ basicAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/EntityId"
        - name: X-Request-Id
          in: header
          required: false
          schema: { type: string, maxLength: 120 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PostPayablePaymentRequest" }
      responses:
        "200":
          description: 已登记付款并返回更新后的发票
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PayableInvoice" }
        "403": { description: 当前角色无权登记付款 }
        "409": { description: 发票版本、请求编号或并发冲突 }
        "422": { description: 付款日期或金额不满足约束 }
  /api/v1/finance/payable-reference-data:
    get:
      operationId: getPayableReferenceData
      summary: 获取采购订单合格收货及其剩余可开票数量
      tags: [Finance]
      security: [{ basicAuth: [] }]
      responses:
        "200":
          description: 可登记应付发票的采购订单
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PayableReferenceData" }
  /api/v1/finance/payable-credit-notes:
    get:
      operationId: listPayableCreditNotes
      summary: 分页查询应付红字发票
      tags: [Finance]
      security: [{ basicAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Query"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: 应付红字发票分页
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PayableCreditNotePage" }
    post:
      operationId: createPayableCreditNote
      summary: 对已过账应付发票开具红字发票
      tags: [Finance]
      security: [{ basicAuth: [] }]
      parameters:
        - name: X-Request-Id
          in: header
          required: false
          schema: { type: string, maxLength: 120 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreatePayableCreditNoteRequest" }
      responses:
        "200":
          description: 已过账红字发票
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PayableCreditNote" }
        "403": { description: 当前角色无权开具红字发票 }
        "409": { description: 请求编号或并发冲突 }
        "422": { description: 日期、数量或原发票状态不满足约束 }
  /api/v1/finance/payable-credit-notes/{id}:
    get:
      operationId: getPayableCreditNote
      summary: 获取应付红字发票详情
      tags: [Finance]
      security: [{ basicAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/EntityId" }]
      responses:
        "200":
          description: 红字发票详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PayableCreditNote" }
        "404": { description: 红字发票不存在或不可见 }
  /api/v1/finance/payable-invoices/{id}/refunds:
    post:
      operationId: postPayableRefund
      summary: 对存在待退余额的应付发票登记从供应商收回的退款
      tags: [Finance]
      security: [{ basicAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/EntityId"
        - name: X-Request-Id
          in: header
          required: false
          schema: { type: string, maxLength: 120 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PostPayableRefundRequest" }
      responses:
        "200":
          description: 已登记退款并返回更新后的发票
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PayableInvoice" }
        "403": { description: 当前角色无权登记退款 }
        "409": { description: 发票版本、请求编号或并发冲突 }
        "422": { description: 退款日期、金额或待退余额不满足约束 }
  /api/v1/finance/payables/payments/{id}/reverse:
    post:
      operationId: reversePayablePayment
      summary: 反核销一笔应付付款或退款
      tags: [Finance]
      security: [{ basicAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/EntityId"
        - name: X-Request-Id
          in: header
          required: false
          schema: { type: string, maxLength: 120 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReversePayablePaymentRequest" }
      responses:
        "200":
          description: 已反核销并返回更新后的发票
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PayableInvoice" }
        "403": { description: 当前角色无权反核销 }
        "409": { description: 请求编号、重复反核销或并发冲突 }
        "422": { description: 日期或原因不满足约束 }
        "404": { description: 付款记录不存在或不可见 }
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: 仅用于当前本地开发身份适配器，正式身份接入后替换。
...
  parameters:
    EntityId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Query:
      name: query
      in: query
      schema: { type: string }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
  schemas:
...
    ReceivableInvoicePage:
      allOf:
        - $ref: "#/components/schemas/PageMetadata"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/ReceivableInvoice" }
    ReceivableInvoice:
      type: object
      required: [id, invoiceNumber, salesOrderId, orderNumber, customerId, customerCode, customerName, currency, invoiceDate, dueDate, taxRate, netAmount, taxAmount, grossAmount, receivedAmount, outstandingAmount, creditBalance, status, version, createdAt, lines, receipts]
      properties:
        id: { type: string, format: uuid }
        invoiceNumber: { type: string }
        salesOrderId: { type: string, format: uuid }
        orderNumber: { type: string }
        customerId: { type: string, format: uuid }
        customerCode: { type: string }
        customerName: { type: string }
        currency: { type: string, minLength: 3, maxLength: 3 }
        invoiceDate: { type: string, format: date }
        dueDate: { type: string, format: date }
        taxRate: { type: number, minimum: 0, maximum: 1 }
        netAmount: { type: number, minimum: 0 }
        taxAmount: { type: number, minimum: 0 }
        grossAmount: { type: number, exclusiveMinimum: 0 }
        receivedAmount: { type: number, minimum: 0 }
        outstandingAmount: { type: number, minimum: 0 }
        creditBalance: { type: number, minimum: 0, description: 红冲后形成的应退客户余额 }
        status: { type: string, enum: [OPEN, PARTIALLY_PAID, PAID, CREDIT_PENDING, SETTLED] }
        version: { type: integer, format: int64, minimum: 0 }
        createdAt: { type: string, format: date-time }
        lines:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/ReceivableInvoiceLine" }
        receipts:
          type: array
          items: { $ref: "#/components/schemas/ReceivableReceipt" }
...
    CreateReceivableInvoiceRequest:
      type: object
      required: [salesOrderId, invoiceDate, dueDate, lines]
      properties:
        salesOrderId: { type: string, format: uuid }
        invoiceDate: { type: string, format: date }
        dueDate: { type: string, format: date }
        lines:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: object
            required: [salesOrderLineId, invoiceQuantity]
            properties:
              salesOrderLineId: { type: string, format: uuid }
              invoiceQuantity: { type: number, minimum: 0.0001 }
    PostReceivableReceiptRequest:
      type: object
      required: [expectedVersion, receiptDate, amount, paymentMethod]
      properties:
        expectedVersion: { type: integer, format: int64, minimum: 0 }
        receiptDate: { type: string, format: date }
        amount: { type: number, minimum: 0.01 }
        paymentMethod: { type: string, enum: [BANK_TRANSFER, CASH, BILL, OTHER] }
        bankReference: { type: [string, "null"], maxLength: 120 }
        note: { type: [string, "null"], maxLength: 500 }
    ReceivableReferenceData:
      type: object
      required: [orders]
      properties:
        orders:
          type: array
          items:
            type: object
            required: [salesOrderId, orderNumber, customerId, customerCode, customerName, currency, taxRate, orderStatus, deliveredAmount, invoicedAmount, remainingAmount, lines]
            properties:
              salesOrderId: { type: string, format: uuid }
              orderNumber: { type: string }
              customerId: { type: string, format: uuid }
              customerCode: { type: string }
              customerName: { type: string }
              currency: { type: string }
              taxRate: { type: number, minimum: 0, maximum: 1 }
              orderStatus: { type: string, enum: [PARTIALLY_SHIPPED, SHIPPED] }
              deliveredAmount: { type: number, minimum: 0 }
              invoicedAmount: { type: number, minimum: 0 }
              remainingAmount: { type: number, minimum: 0 }
              lines:
                type: array
                items: { $ref: "#/components/schemas/ReceivableReferenceLine" }
...
    PayableInvoicePage:
      allOf:
        - $ref: "#/components/schemas/PageMetadata"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/PayableInvoice" }
    PayableInvoice:
      type: object
      required: [id, invoiceNumber, supplierInvoiceNumber, purchaseOrderId, orderNumber, supplierId, supplierCode, supplierName, currency, invoiceDate, dueDate, taxRate, netAmount, taxAmount, grossAmount, paidAmount, outstandingAmount, creditBalance, status, version, createdAt, lines, payments]
      properties:
        id: { type: string, format: uuid }
        invoiceNumber: { type: string }
        supplierInvoiceNumber: { type: string }
        purchaseOrderId: { type: string, format: uuid }
        orderNumber: { type: string }
        supplierId: { type: string, format: uuid }
        supplierCode: { type: string }
        supplierName: { type: string }
        currency: { type: string, minLength: 3, maxLength: 3 }
        invoiceDate: { type: string, format: date }
        dueDate: { type: string, format: date }
        taxRate: { type: number, minimum: 0, maximum: 1 }
        netAmount: { type: number, minimum: 0 }
        taxAmount: { type: number, minimum: 0 }
        grossAmount: { type: number, exclusiveMinimum: 0 }
        paidAmount: { type: number, minimum: 0 }
        outstandingAmount: { type: number, minimum: 0 }
        creditBalance: { type: number, minimum: 0, description: 红冲后形成的应向供应商收回余额 }
        status: { type: string, enum: [OPEN, PARTIALLY_PAID, PAID, CREDIT_PENDING, SETTLED] }
        version: { type: integer, format: int64, minimum: 0 }
        createdAt: { type: string, format: date-time }
        lines:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/PayableInvoiceLine" }
        payments:
          type: array
          items: { $ref: "#/components/schemas/PayablePayment" }
...
    CreatePayableInvoiceRequest:
      type: object
      required: [purchaseOrderId, supplierInvoiceNumber, invoiceDate, dueDate, lines]
      properties:
        purchaseOrderId: { type: string, format: uuid }
        supplierInvoiceNumber: { type: string, minLength: 1, maxLength: 80 }
        invoiceDate: { type: string, format: date }
        dueDate: { type: string, format: date }
        lines:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: object
            required: [purchaseOrderLineId, invoiceQuantity]
            properties:
              purchaseOrderLineId: { type: string, format: uuid }
              invoiceQuantity: { type: number, minimum: 0.0001 }
    PostPayablePaymentRequest:
      type: object
      required: [expectedVersion, paymentDate, amount, paymentMethod]
      properties:
        expectedVersion: { type: integer, format: int64, minimum: 0 }
        paymentDate: { type: string, format: date }
        amount: { type: number, minimum: 0.01 }
        paymentMethod: { type: string, enum: [BANK_TRANSFER, CASH, BILL, OTHER] }
        bankReference: { type: [string, "null"], maxLength: 120 }
        note: { type: [string, "null"], maxLength: 500 }
    PayableReferenceData:
      type: object
      required: [orders]
      properties:
        orders:
          type: array
          items:
            type: object
            required: [purchaseOrderId, orderNumber, supplierId, supplierCode, supplierName, currency, taxRate, orderStatus, acceptedAmount, invoicedAmount, remainingAmount, lines]
            properties:
              purchaseOrderId: { type: string, format: uuid }
              orderNumber: { type: string }
              supplierId: { type: string, format: uuid }
              supplierCode: { type: string }
              supplierName: { type: string }
              currency: { type: string }
              taxRate: { type: number, minimum: 0, maximum: 1 }
              orderStatus: { type: string }
              acceptedAmount: { type: number, minimum: 0 }
              invoicedAmount: { type: number, minimum: 0 }
              remainingAmount: { type: number, minimum: 0 }
              lines:
                type: array
                items: { $ref: "#/components/schemas/PayableReferenceLine" }
...
    ReceivableCreditNotePage:
      allOf:
        - $ref: "#/components/schemas/PageMetadata"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/ReceivableCreditNote" }
    ReceivableCreditNote:
      type: object
      required: [id, creditNoteNumber, originalInvoiceId, originalInvoiceNumber, salesOrderId, orderNumber, customerId, customerCode, customerName, currency, creditNoteDate, dueDate, taxRate, netAmount, taxAmount, grossAmount, reason, status, version, createdAt, lines]
      properties:
        id: { type: string, format: uuid }
        creditNoteNumber: { type: string }
        originalInvoiceId: { type: string, format: uuid }
        originalInvoiceNumber: { type: string }
        salesOrderId: { type: string, format: uuid }
        orderNumber: { type: string }
        customerId: { type: string, format: uuid }
        customerCode: { type: string }
        customerName: { type: string }
        currency: { type: string, minLength: 3, maxLength: 3 }
        taxNoticeNumber: { type: [string, "null"], maxLength: 80, description: 增值税红字信息表编号，可空 }
        creditNoteDate: { type: string, format: date }
        dueDate: { type: string, format: date }
        taxRate: { type: number, minimum: 0, maximum: 1 }
        netAmount: { type: number, maximum: 0, description: 负数净额 }
        taxAmount: { type: number, maximum: 0, description: 负数税额 }
        grossAmount: { type: number, exclusiveMaximum: 0, description: 负数含税金额 }
        reason: { type: string, minLength: 4, maxLength: 500 }
        status: { type: string, enum: [POSTED] }
        version: { type: integer, format: int64, minimum: 0 }
        createdAt: { type: string, format: date-time }
        lines:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/ReceivableCreditNoteLine" }
...
    CreateReceivableCreditNoteRequest:
      type: object
      required: [originalInvoiceId, creditNoteDate, dueDate, reason, lines]
      properties:
        originalInvoiceId: { type: string, format: uuid }
        taxNoticeNumber: { type: [string, "null"], maxLength: 80 }
        creditNoteDate: { type: string, format: date }
        dueDate: { type: string, format: date }
        reason: { type: string, minLength: 4, maxLength: 500 }
        lines:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: object
            required: [originalInvoiceLineId, creditQuantity]
            properties:
              originalInvoiceLineId: { type: string, format: uuid }
              creditQuantity: { type: number, minimum: 0.0001 }
              unitPrice: { type: [number, "null"], minimum: 0, description: 不填则取原蓝字发票行单价；用于折让场景 }
    PostReceivableRefundRequest:
      type: object
      required: [expectedVersion, refundDate, amount, paymentMethod]
      properties:
        expectedVersion: { type: integer, format: int64, minimum: 0 }
        refundDate: { type: string, format: date }
        amount: { type: number, minimum: 0.01 }
        paymentMethod: { type: string, enum: [BANK_TRANSFER, CASH, BILL, OTHER] }
        bankReference: { type: [string, "null"], maxLength: 120 }
        note: { type: [string, "null"], maxLength: 500 }
    ReverseReceivableReceiptRequest:
      type: object
      required: [reversalDate, reason]
      properties:
        reversalDate: { type: string, format: date }
        reason: { type: string, minLength: 4, maxLength: 500 }
    PayableCreditNotePage:
      allOf:
        - $ref: "#/components/schemas/PageMetadata"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/PayableCreditNote" }
...
    CreatePayableCreditNoteRequest:
      type: object
      required: [originalInvoiceId, creditNoteDate, dueDate, reason, lines]
      properties:
        originalInvoiceId: { type: string, format: uuid }
        supplierCreditNoteNumber: { type: [string, "null"], maxLength: 80 }
        taxNoticeNumber: { type: [string, "null"], maxLength: 80 }
        creditNoteDate: { type: string, format: date }
        dueDate: { type: string, format: date }
        reason: { type: string, minLength: 4, maxLength: 500 }
        lines:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: object
            required: [originalInvoiceLineId, creditQuantity]
            properties:
              originalInvoiceLineId: { type: string, format: uuid }
              creditQuantity: { type: number, minimum: 0.0001 }
              unitPrice: { type: [number, "null"], minimum: 0 }
...